Change font API again
We shouldn't try to hide that text is getting rendered from a font
atlas with a certain size. Users have to create that atlas anyway,
so they know what sizes will work well and which won't. The idea
to let users pass a size for text rendering is nice, but it doesn't
work out with an atlas rendering. It would work with fancy rendering
e.g. Dobbie or Slug, but we can't do that in a platform-independent
way so let's stick with TextureFont for now.